Move type to its own file.  Add contracts.
