+++ src/Features/Core/Portable/AddImport/AddImportFixKind.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
+
+namespace Microsoft.CodeAnalysis.AddImport
+{
+    internal enum AddImportFixKind
+    {
+        ProjectSymbol,
+        MetadataSymbol,
+        PackageSymbol,
+        ReferenceAssemblySymbol,
+    }
+}
