+++ src/Features/Core/Portable/AddImport/CodeActions/MetadataSymbolReferenceCodeAction.cs
@@ -12,6 +12,7 @@ namespace Microsoft.CodeAnalysis.AddImport
             public MetadataSymbolReferenceCodeAction(Document originalDocument, AddImportFixData fixData)
                 : base(originalDocument, fixData)
             {
+                Contract.ThrowIfFalse(fixData.Kind == AddImportFixKind.MetadataSymbol);
             }
 
             protected override Project UpdateProject(Project project)
