+++ src/Features/Core/Portable/AddImport/CodeActions/ProjectSymbolReferenceCodeAction.cs
@@ -1,5 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using Roslyn.Utilities;
+
 namespace Microsoft.CodeAnalysis.AddImport
 {
     internal abstract partial class AbstractAddImportCodeFixProvider<TSimpleNameSyntax>
@@ -17,6 +19,7 @@ namespace Microsoft.CodeAnalysis.AddImport
                 AddImportFixData fixData)
                 : base(originalDocument, fixData)
             {
+                Contract.ThrowIfFalse(fixData.Kind == AddImportFixKind.ProjectSymbol);
             }
 
             private bool ShouldAddProjectReference()
